Calculating daily averages for week/months Trends

+++ app/src/main/java/dev/gaborbiro/dailymacros/features/trends/model/TrendsViewState.kt
@@ -11,4 +11,4 @@ enum class TimeScale { DAYS, WEEKS, MONTHS }
 
 data class MacroChartData(val datasets: List<MacroChartDataset>)
 data class MacroChartDataset(val name: String, val color: Color, val set: List<MacroChartDataPoint>, val now: MacroChartDataPoint?)
-data class MacroChartDataPoint(val index: Int, val label: String, val value: Float?)
+data class MacroChartDataPoint(val index: Int, val label: String, val value: Double?)
